+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "client",
-  "homepage": "http://thirtythreeboas.github.io/swift-learner",
+  "homepage": "http://thirtythreeboas.github.io/",
   "version": "0.1.0",
   "private": true,
   "dependencies": {
