+++ src/components/Details/TestSettings.module.scss
@@ -13,6 +13,7 @@
 .details {
   display: flex;
   flex-direction: column;
+  // width: 360px;
   width: 360px;
 }
