+++ src/components/Result/Result.module.scss
@@ -18,7 +18,6 @@
 
 .resultData {
   width: 100%;
-  header {
     .headerResult {
       display: flex;
       border-bottom: 0;
@@ -26,7 +25,6 @@
         box-shadow: none;
       }
     }
-  }
 }
 
 .resultsRow {
