+++ src/components/TestLogic/TestLogic.module.scss
@@ -3,57 +3,3 @@
   display: flex;
   justify-content: center;
 }
-
-.inputContainer {
-  display: flex;
-  gap: 10px;
-}
-
-.inputBlock {
-  display: flex;
-  flex-direction: column;
-  gap: 10px;
-}
-
-#word-to-translate {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  min-width: 200px;
-  height: 30px;
-  border: 1px solid #9b8686;
-  border-radius: 4px;
-}
-
-.word-to-enter {
-  text-align: center;
-  min-width: 200px;
-  height: 30px;
-  border: 1px solid #9b8686;
-  border-radius: 4px;
-  padding: 0;
-}
-
-.inputWordButton {
-  width: 100px;
-  background-color: #7eff4f;
-  // border: 1px solid #321a1a;
-  font-size: 20px;
-  border: none;
-  border-radius: 3px;
-  cursor: pointer;
-  &:hover {
-    background-color: #00ff27;
-  }
-}
-
-.amountBadges {
-  position: absolute;
-  top: 0;
-  right: 0;
-  display: flex;
-  gap: 10px;
-  span {
-    display: block;
-  }
-}
