+++ src/components/UI/Loading/index.js
@@ -1,0 +1,1 @@
+export * from './Loading'
