+++ src/components/UI/SwitchLangs/SwitchLangs.module.scss
