+++ src/components/UI/SwitchLangs/index.js
@@ -1,0 +1,1 @@
+export * from './SwitchLangs'
