+++ src/components/WelcomeImg/WelcomeImg.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import styles from './WelcomeImg.module.scss'
+
+export const WelcomeImg = () => {
+  return (
+    <img
+      className={styles.welcomeImg}
+      src="https://sun9-50.userapi.com/impg/YgpXpztaf-cu_ZzTQcu4ea4aQBFEacRTnOzRdA/h5m1xJ1ycAk.jpg?size=654x435&quality=96&sign=90681a59c19259c5f2abc3658d21d1d3&type=album"
+      alt="img"
+    />
+  )
+}
