+++ src/components/WelcomeImg/WelcomeImg.module.scss
@@ -1,0 +1,3 @@
+.welcomeImg {
+  
+}
